third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/scripts-panel-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that scripts panel UI elements work as intended.


Running: testInitialLoad
Navigator:
top
  example.com
    (index)
  (no domain)
    bar.js
    baz.js
    foo.js
    foobar.js

Running: testReset
Navigator:
top
  example.com
    (index)
  (no domain)
    bar.js
    baz.js
    foo.js
Revealing in navigator.
Navigator:
top
  example.com
    (index)
  (no domain)
    bar.js
    baz.js
    foo.js
Navigator:
top
  example.com
    (index)
Navigator:
top
  example.com
    (index)
  (no domain)
    bar.js

Running: testDebuggerUISourceCodeAddedAndRemoved
Navigator:
top
  example.com
    (index)
  (no domain)
    foo.js
Navigator:
top
  example.com
    (index)
  (no domain)
    foo.js
    source.js
